Group search functionality together.
Enable persistent "very magic" search so that almost all symbols have
special meaning, and need to be escaped if you want them to match
literally.

Precede / or ? with <leader> to do a non-very-magic search to
search for literal characters.

In a normal search, just delete the /v if you want to do a normal
magic/nomagic search.